DAT AA
#1

+++ Chocobro/Job Modules/Job.cs
@@ -16,8 +16,7 @@ namespace Chocobro {
     public int PIE { get; set; }
     public int WEP { get; set; }
     public double AADMG = 0.0;
-    
-    public int AAPOT = 0;
+    public double AAPOT = 0;
     public double CRIT = 341;
     public int DTR = 202;
     public int ACC = 341;
